+++ src/arguments/index.ts
@@ -1,13 +1,17 @@
+import { dofusCoordinates } from './dofus/coordinates';
 import { dofusDimension } from './dofus/dimension';
 import { dofusJob } from './dofus/job';
 import { dofusServer } from './dofus/server';
 import { duration } from './time/duration';
 import { line } from './text/line';
+import { oneline } from './text/oneline';
 
 export const argumentTypes = {
+  dofusCoordinates,
   dofusDimension,
   dofusJob,
   dofusServer,
   duration,
-  line
+  line,
+  oneline
 };