+++ src/constants/dofus/zaaps.ts
@@ -1,0 +1,587 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+
+export const ZAAPS = [
+  {
+    zone: 'cemetery',
+    x: 3,
+    y: 0,
+    world: 'incarnam',
+    type: 'zaap'
+  },
+  {
+    zone: 'pastures',
+    x: 2,
+    y: -5,
+    world: 'incarnam',
+    type: 'zaap'
+  },
+  {
+    zone: 'way_of_souls',
+    x: -1,
+    y: -3,
+    world: 'incarnam',
+    type: 'zaap'
+  },
+  {
+    zone: 'bonta_city_center',
+    x: -32,
+    y: -56,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'kanig_village',
+    x: 0,
+    y: -56,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'heroes_cemetery',
+    x: -21,
+    y: -58,
+    world: 'main',
+    type: 'stagecoach'
+  },
+  {
+    zone: 'cania_fields',
+    x: -31,
+    y: -48,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'cania_fields',
+    x: -27,
+    y: -36,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'rocky_plains',
+    x: -17,
+    y: -47,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'cania_lake',
+    x: -3,
+    y: -42,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'cradle',
+    x: 1,
+    y: -32,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'cania_massif',
+    x: -13,
+    y: -28,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'rocky_roads',
+    x: -20,
+    y: -20,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'lousy_pig_plain',
+    x: -5,
+    y: -23,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'astrub_city',
+    x: 5,
+    y: -18,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'cania_bay',
+    x: -29,
+    y: -12,
+    world: 'main',
+    type: 'stagecoach'
+  },
+  {
+    zone: 'dopple_village',
+    x: -16,
+    y: 1,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'breeder_village',
+    x: -34,
+    y: -8,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'caravan_alley',
+    x: -25,
+    y: 12,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'desecrated_highlands',
+    x: -15,
+    y: 25,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'fungus_domain',
+    x: -12,
+    y: 29,
+    world: 'main',
+    type: 'stagecoach'
+  },
+  {
+    zone: 'brakmar_city_walls',
+    x: -26,
+    y: 29,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'brakmar_city_center',
+    x: -26,
+    y: 35,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'scaraleaf_plain',
+    x: -1,
+    y: 24,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'dreggon_peninsula',
+    x: 11,
+    y: 30,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'sufokia',
+    x: 13,
+    y: 26,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'sufokian_shoreline',
+    x: 10,
+    y: 22,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'jelly_peninsula',
+    x: -3,
+    y: 27,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'edge_of_the_evil_forest',
+    x: -1,
+    y: 13,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'gobball_corner',
+    x: 5,
+    y: 7,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'amakna_village',
+    x: -2,
+    y: 0,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'madrestam_harbour',
+    x: 7,
+    y: -4,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'amakna_castle',
+    x: 3,
+    y: -5,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'crackler_mountain',
+    x: -5,
+    y: -8,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'kawaii_river',
+    x: 6,
+    y: -1,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'low_crackler_mountain',
+    x: -1,
+    y: 8,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'coastal_village',
+    x: -46,
+    y: 18,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'grassy_plains',
+    x: -57,
+    y: 4,
+    world: 'main',
+    type: 'scraplanes'
+  },
+  {
+    zone: 'grassy_plains',
+    x: -49,
+    y: 14,
+    world: 'main',
+    type: 'scraplanes'
+  },
+  {
+    zone: 'grassy_plains',
+    x: -56,
+    y: 22,
+    world: 'main',
+    type: 'scraplanes'
+  },
+  {
+    zone: 'coral_beach',
+    x: -57,
+    y: 1,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'zoth_village',
+    x: -53,
+    y: 18,
+    world: 'hakam_tree',
+    type: 'zaap'
+  },
+  {
+    zone: 'canopy_village',
+    x: -54,
+    y: 16,
+    world: 'hakam_tree',
+    type: 'zaap'
+  },
+  {
+    zone: 'canopy_village',
+    x: -54,
+    y: -4,
+    world: 'hakam_tree',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'vulkania_village',
+    x: -48,
+    y: 44,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'minotoror_island',
+    x: -43,
+    y: -16,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'minotoror_island',
+    x: -41,
+    y: -18,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'alliance_temple',
+    x: 14,
+    y: 36,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'alliance_temple',
+    x: 13,
+    y: 35,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'turtle_beach',
+    x: 35,
+    y: 12,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'turtle_beach',
+    x: 35,
+    y: 4,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'cawwot_island',
+    x: 24,
+    y: -2,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'cawwot_island',
+    x: 25,
+    y: -4,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'cawwot_island',
+    x: 23,
+    y: -3,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'abandoned_labowatowies',
+    x: 27,
+    y: -14,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'sarakech_port',
+    x: 14,
+    y: -55,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'sarakech_port',
+    x: 16,
+    y: -57,
+    world: 'main',
+    type: 'sand_yacht'
+  },
+  {
+    zone: 'dunes_of_bones',
+    x: 15,
+    y: -58,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'castuc_territory',
+    x: 12,
+    y: -63,
+    world: 'main',
+    type: 'sand_yacht'
+  },
+  {
+    zone: 'gorge_of_howling_winds',
+    x: 11,
+    y: -70,
+    world: 'main',
+    type: 'sand_yacht'
+  },
+  {
+    zone: 'pandala_village',
+    x: 20,
+    y: -29,
+    world: 'main',
+    type: 'sand_yacht'
+  },
+  {
+    zone: 'aerdala',
+    x: 21,
+    y: -37,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'nolifis_cemetery',
+    x: 34,
+    y: -40,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'frigost_village',
+    x: -78,
+    y: -41,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'frigost_village',
+    x: -76,
+    y: -46,
+    world: 'main',
+    type: 'frigostian_ski'
+  },
+  {
+    zone: 'icefields',
+    x: -68,
+    y: -34,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'icefields',
+    x: -70,
+    y: -33,
+    world: 'main',
+    type: 'imp_carrier'
+  },
+  {
+    zone: 'icefields',
+    x: -84,
+    y: -49,
+    world: 'main',
+    type: 'frigostian_ski'
+  },
+  {
+    zone: 'frozen_lake',
+    x: -64,
+    y: -55,
+    world: 'main',
+    type: 'frigostian_ski'
+  },
+  {
+    zone: 'almas_cradle',
+    x: -56,
+    y: -74,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'tears_of_ouronigride',
+    x: -69,
+    y: -87,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'asparah_gorge',
+    x: -78,
+    y: -82,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'petrified_forest',
+    x: -76,
+    y: -66,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'snowbound_village',
+    x: -77,
+    y: -73,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'fangs_of_glass',
+    x: -66,
+    y: -65,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'mount_scauldron',
+    x: -63,
+    y: -71,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'entrance_to_harebourgs_castle',
+    x: -67,
+    y: -75,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'sakai_harbour',
+    x: -51,
+    y: -38,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'sakai_plain',
+    x: -55,
+    y: -41,
+    world: 'main',
+    type: 'frigostian_carrier'
+  },
+  {
+    zone: 'albatrocious_rock',
+    x: -83,
+    y: -58,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'kwismas_haven',
+    x: -30,
+    y: -84,
+    world: 'main',
+    type: 'boat'
+  },
+  {
+    zone: 'imp_village',
+    x: -16,
+    y: -24,
+    world: 'main',
+    type: 'zaap'
+  },
+  {
+    zone: 'crocuzko',
+    x: -83,
+    y: -15,
+    world: 'main',
+    type: 'zaap'
+  }
+];
+
+/* eslint-enable @typescript-eslint/naming-convention */