+++ CHANGELOG.md
@@ -7,8 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-<<<<<<< HEAD
-=======
+## [0.2.0] - 2021-02-21
+
+### Added
+
+- Command `zaap`: find the closest tramports to a coordinate
+- Argument `dofusCoordinates`: parse Dofus maps' coordinates
+- Tests: `zaap`command and `dofusCoordinates` type
+- Providers: new provider for tracking issues
+- Commands `issue`: report bugs, ask for features and track tasks
+- Arguments `oneline`: match the first line of a message only
+- Utils: pad a string with repeating characaters
+
+### Changed
+
+- Command `portal`: Display closest transports to the portal
+- Providers `mongoose-cached`: find the first document that satisfies requirements
+- Commands `help`: do not list never-matching arguments in the usage
+
+### Fixed
+
+- Command `job`: fetch non-cached members from Discord, ignore missing members and remove duplicates
+- Command `locale`: display the new language in the correct language
+>>>>>>> main
 ## [0.1.7] - 2021-02-15
 
 ### Fixed
@@ -119,7 +140,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Task `almanax`: automatically fetch the [almanax](http://www.krosmoz.com/en/almanax) of the day at midnight
 - Task `twitter`: fetch news from twitter in (almost) real-time
 
-[unreleased]: https://github.com/brinkflew/krosmobot/compare/v0.1.7...HEAD
+[unreleased]: https://github.com/brinkflew/krosmobot/compare/v0.2.0...HEAD
+[0.1.7]: https://github.com/brinkflew/krosmobot/compare/v0.1.7...v0.2.0
 [0.1.7]: https://github.com/brinkflew/krosmobot/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/brinkflew/krosmobot/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/brinkflew/krosmobot/compare/v0.1.4...v0.1.5